+++ backend/app.js
@@ -4,8 +4,11 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
+// import routes
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
+var weather = require('./routes/weather');
+//var sendText = require('./routes/sms');
 
 var app = express();
 
@@ -19,8 +22,11 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
+// Specify routers
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/weather', weather);
+//app.use('/sendText', sendText);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
