+++ backend/package.json
@@ -6,9 +6,9 @@
     "start": "nodemon ./bin/www"
   },
   "dependencies": {
-    "axios": "^1.1.3",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
+    "dotenv": "^16.0.3",
     "express": "~4.16.1",
     "hbs": "~4.0.4",
     "http-errors": "~1.6.3",
