+++ backend/routes/index.js
@@ -1,45 +1,26 @@
 var express = require('express');
 var router = express.Router();
-
-
-
-
-module.exports = router;
-
-const axios = require("axios");
-
-const options = {
-  method: 'GET',
-  url: 'http://api.weatherapi.com/v1/current.json',
-  params: {q: 'London', dt: '2022-12-25', key: '38556edbb21c4d2ea89231336221810'},
-  headers: {
-    'X-RapidAPI-Key': '38556edbb21c4d2ea89231336221810',
-    'X-RapidAPI-Host': 'weatherapi-com.p.rapidapi.com'
-  }
-};
-
-const accountSid = ' '; // Your Account SID from www.twilio.com/console
-const authToken = ' '; // Your Auth Token from www.twilio.com/console
 const twilio = require('twilio');
+require('dotenv').config();
+
+
+const accountSid = process.env.TWILIO_ACCOUNT_SID; // Your Account SID from www.twilio.com/console
+const authToken = process.env.AUTHENTICATION_TOKEN; // Your Auth Token from www.twilio.com/console
 
 const client = new twilio(accountSid, authToken);
 
 client.messages
   .create({
     body: 'Hello from Node',
-    to: '+', // Text this number
+    to: '+', // Text to this number
     from: '+', // From a valid Twilio number
-  })
-  .then((message) => console.log(message.sid));
+  }).then((message) => console.log('Send Success'));
 
-
-
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  axios.request(options).then(function (response) {
-    res.render('weather',  {response:response.data}  );
-  }).catch(function (error) {
-    console.error(error);
+router.get('/', function(req, res){
+  res.render('sample', {
+    title: 'Express',
+    subTitle: 'subExpress'
   });
-  
 });
+
+module.exports = router;
