+++ backend/routes/test.js
@@ -1,0 +1,2 @@
+require('dotenv').config()
+console.log(process.env.AUTHENTICATION_TOKEN);
