+++ backend/views/users.hbs
@@ -1,0 +1,6 @@
+<img src= {{response.picture.large}} width="250" height="250" >
+<h1>Name: {{response.name.first}} {{response.name.last}} </h1>
+<h2>Age: {{response.dob.age}} </h2>
+<h2>Location: {{response.location.city}}, {{response.location.country}}</h2>
+<h2>Email: {{response.email}} </h2>
+<h2>Phone: {{response.phone}} </h2>
