+++ backend/views/weather.hbs
@@ -1,8 +1,8 @@
 <h1>{{response.location.name}}</h1>
 <h1>{{response.location.region}}</h1>
+<h2>{{response.location.localtime}}</h2>
 <h1>{{response.current.temp_f}}f</h1>
 <h1>{{response.current.temp_f}}f</h1>
-<h2></h2>
